Added classifier, problems: 1) classifier accuracy during training is 1.0, loss 0.0001, during evaluation - 1.0/0.0001. during testing - accuracy is 0.0 2) tracknet precision and recall are 0.8, not 0.95 3) very slow preprocessing

+++ ariadne/lightning.py
@@ -15,7 +15,7 @@ class TrainModel(pl.LightningModule):
                  data_loader: BaseDataLoader.__class__):
         super().__init__()
         # configure model and criterion
-        self.model = model()
+        self.model = model().float()
         self.criterion = criterion()
         self.optimizer = optimizer
         self.metrics = metrics
@@ -36,9 +36,12 @@ class TrainModel(pl.LightningModule):
         return metric_vals
 
     def _forward_batch(self, batch, val=False):
+        #print(batch)
         x, y = batch
+        #print('in batch: ', y)
         y_pred = self.model(**x)
-        loss = self.criterion(y_pred, y)
+        #print('preds: ', y_pred)
+        loss = self.criterion(y_pred, y.long())
         metric_vals = self._calc_metrics(y_pred, y)
         return {'loss': loss, **metric_vals}
 
@@ -46,6 +49,8 @@ class TrainModel(pl.LightningModule):
         result_dict = self._forward_batch(batch)
         tqdm_dict = {f'train_{k}': v for k, v in result_dict.items()}
         result = pl.TrainResult(result_dict['loss'])
+        #print(result)
+        #result['minimize'].backward()
         result.log_dict(tqdm_dict, on_step=False, prog_bar=True)
         return result
